flask-gateway/requirements.txt
click==8.0.3
Flask==2.0.2
Flask-Cors==3.0.10
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
Werkzeug==2.0.2
ldap3==2.9.1
columnar==1.4.1
GitPython>=3.1.18
pyjwt[crypto]
flask-apscheduler
cryptography
gunicorn==20.1.0
openpyxl
prometheus-client
prometheus-flask-exporter
psutil
kubernetes
fred-metadata
